'Refactored by Sourcery'

+++ app/tasks.py
@@ -82,24 +82,26 @@ except:
     logging.critical("Error in logging config, ignoring")
 
 # Migrate from Plex to Global Settings:
-if Settings.select().where(Settings.key == 'admin_username').exists():
-    if Settings.select().where(Settings.key == 'plex_verified').exists():
-        if not Settings.select().where(Settings.key == 'server_type').exists() :
-            try:
-                os.system("cp ./database/database.db ./database/1.6.5-database-backup.db")
-                logging.info("Database backup created due to major version update.")
-            except:
-                pass
-            Settings.create(key='server_type', value='plex')
-            Settings.create(key='api_key', value=Settings.get(Settings.key == 'plex_token').value)
-            Settings.delete().where(Settings.key == 'plex_token').execute()
-            Settings.create(key='server_url', value=Settings.get(Settings.key == 'plex_url').value)
-            Settings.delete().where(Settings.key == 'plex_url').execute()
-            Settings.create(key='server_name', value=Settings.get(Settings.key == 'plex_name').value)
-            Settings.delete().where(Settings.key == 'plex_name').execute()
-            Settings.create(key='libraries', value=Settings.get(Settings.key == 'plex_libraries').value)
-            Settings.delete().where(Settings.key == 'plex_libraries').execute()
-            Settings.create(key='server_verified', value=Settings.get(Settings.key == 'plex_verified').value)
-            Settings.delete().where(Settings.key == 'plex_verified').execute()
+if (
+    Settings.select().where(Settings.key == 'admin_username').exists()
+    and Settings.select().where(Settings.key == 'plex_verified').exists()
+    and not Settings.select().where(Settings.key == 'server_type').exists()
+):
+    try:
+        os.system("cp ./database/database.db ./database/1.6.5-database-backup.db")
+        logging.info("Database backup created due to major version update.")
+    except:
+        pass
+    Settings.create(key='server_type', value='plex')
+    Settings.create(key='api_key', value=Settings.get(Settings.key == 'plex_token').value)
+    Settings.delete().where(Settings.key == 'plex_token').execute()
+    Settings.create(key='server_url', value=Settings.get(Settings.key == 'plex_url').value)
+    Settings.delete().where(Settings.key == 'plex_url').execute()
+    Settings.create(key='server_name', value=Settings.get(Settings.key == 'plex_name').value)
+    Settings.delete().where(Settings.key == 'plex_name').execute()
+    Settings.create(key='libraries', value=Settings.get(Settings.key == 'plex_libraries').value)
+    Settings.delete().where(Settings.key == 'plex_libraries').execute()
+    Settings.create(key='server_verified', value=Settings.get(Settings.key == 'plex_verified').value)
+    Settings.delete().where(Settings.key == 'plex_verified').execute()
